WIP Add Artsakh
Removed from AUG in
https://github.com/anki-geo/ultimate-geography/pull/620 and copied
here verbatim.

Artsakh may no longer be a sovereign state, and the note needs
updating (past tense and date range) but the information itself and
hence the note, might still be of interest to the more
historically-minded.

Assuming that we do actually add Artsakh (with updates to the note),
we might also want a UG::Historical (naming?) tag.